media-source/MediaSource-endOfStream_readyState.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>readyState attribute not in the "open" state</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(function() {
    var mediaSource = new MediaSource();
    assert_throws({name : 'InvalidStateError'},
                  function() {
                      mediaSource.endOfStream();
                  });
}, 'MediaSource#endOfStream : If the readyState attribute is not in the "open" state then throw an INVALID_STATE_ERR exception and abort these steps.');
</script>
</body>
</html>
